chore: move effect tests to ExperimentalSuite
